fix: update osmosis lp-stake reference tx
This updates the osmosis lp-stake reference tx data to match the only
actual on-chain staking TX from the reference seed. Unfortunately, it
was signed with LEGACY_AMINO_JSON, so the signature doesn't match this
library's output... but the on-chain signature has been manually
verified to match the LEGACY_AMINO_JSON encoding of the corrected
reference TX input.

Note that there's a serialization bug with osmosis lp-stake txes, which
means that the freshly-updated output for this reference tx is
definitely still wrong... but at least it's *less* wrong. Sorta.