fix: a couple osmosis encoding bugs (#12)
* chore: whitespace fixes

* chore: normalize thorchain reference txes

* fix: update bad osmosis reference txids

* fix: update osmosis lp-stake reference tx

This updates the osmosis lp-stake reference tx data to match the only
actual on-chain staking TX from the reference seed. Unfortunately, it
was signed with LEGACY_AMINO_JSON, so the signature doesn't match this
library's output... but the on-chain signature has been manually
verified to match the LEGACY_AMINO_JSON encoding of the corrected
reference TX input.

Note that there's a serialization bug with osmosis lp-stake txes, which
means that the freshly-updated output for this reference tx is
definitely still wrong... but at least it's *less* wrong. Sorta.

* fix: don't skip "0" amounts, it breaks some reference txes

The previous behavior prevented the reference TX serializations for
osmosis lp-add and undelegate from matching the on-chain data. Note
that osmosis lp-stake didn't match on-chain data before, and it still
doesn't, both due to a serialization bug and because the reference TX
doesn't use SIGN_MODE_DIRECT.

* chore: expand reference txids to full on-chain data

The data here is formatted as shown in Mintscan's JSON view.

* fix: osmosis tx encoding bugs

The serialized reference TX for osmosis lp-stake is now correct, but the
signature still doesn't match what's on-chain because the on-chain
version of the transaction was signed with LEGACY_AMINO_JSON.

* chore: add new osmosis reference transactions

* cleanup and fix tests

* chore: version bump

Co-authored-by: kaladinlight <35275952+kaladinlight@users.noreply.github.com>
Co-authored-by: pastaghost <pastaghost@2520.io>